requirements.txt
bleach==1.5.0
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
dill==0.2.7.1
Django==1.11.7
enum34==1.1.6
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
google-api-core==0.1.1
google-auth==1.2.0
google-cloud==0.30.0
google-cloud-bigquery==0.28.0
google-cloud-bigtable==0.28.1
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.0
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.0
google-cloud-pubsub==0.29.0
google-cloud-resource-manager==0.28.0
google-cloud-runtimeconfig==0.28.0
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.16.0
google-cloud-translate==1.3.0
google-cloud-videointelligence==0.28.0
google-cloud-vision==0.28.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.7.0
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
image==1.5.16
Markdown==2.6.9
matplotlib==2.1.0
numpy==1.13.3
oauth2client==3.0.0
olefile==0.44
opencv-python==3.3.0.10
Pillow==4.3.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.4.0
psutil==5.4.1
pyasn1==0.3.7
pyasn1-modules==0.1.5
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
urllib3==1.22
utils==0.9.0
Werkzeug==0.12.2
